% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_y_axis_labels.r
\name{make_y_axis_labels}
\alias{make_y_axis_labels}
\title{Create a function for making y-axis labels for regression graphs}
\usage{
make_y_axis_labels(label_dimension_lookup, dp_text_palette)
}
\arguments{
\item{label_dimension_lookup}{A lookup created using the prepare_figure_data
function.}

\item{dp_text_palette}{A color palette assigning color to dimension to use
when dimensions are colored by text.}
}
\value{
A function for creating y-axis labels to use for regression graphs.
}
\description{
This function returns a function for making y-axis labels using a palette for
text colors and label dimension lookup from prepare_figure_data(). The labels
are colored by dimension, and the indicator half of the combined label is
bold.
}
\details{
The function takes label_dimension_lookup and dp_text_palette to
assign palette colors to each dimension. It splits the combined label back
into indicator and metric to bold the indicator then recombines them. It
pastes the newly formatted label and color assignment together.
}
\examples{
\dontrun{
# Sample data frame
df <- data.frame(
  dimension = c("d1", "d2", "d3", "d4"),
  indicator = c("earth", "wind", "fire", "music"),
  metric = c("soil", "air", "heat", "notes"),
  combined_label = c("earth - soil", "wind - air", "fire - heat", "music - notes"),
  estimate = c(0.01, 0.05, -0.2, 0.07)
  )

# Use make_y_axis_labels when plotting
ggplot(data = df, mapping = aes(x = estimate, y = combined_label)) +
  geom_point() +
  geom_errorbarh() +
  geom_vline(xintercept = 0) +
  scale_y_discrete(
    labels = make_y_axis_labels(label_dimension_lookup, dp_text_palette))
}

}
