% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_meta.R
\docType{data}
\name{dp_meta}
\alias{dp_meta}
\title{Data Paper Metadata}
\format{
\subsection{\code{data_paper_meta} A data frame in long format with one value per}{

row.
\describe{
\item{quality}{Qualitative rating of metric quality. 3 is best, 1 is worst.}
\item{dimension}{Framework dimension}
\item{index}{Framework index}
\item{indicator}{Framework indicator}
\item{indicator type}{One of four categories: target, practice, result,
outcome.}
\item{keep}{If it stays in the revised metric set, it gets an \code{x}}
\item{analyze}{If we can reasonable assign a normative value to it (i.e.,
if we can say which direction is good or bad), it gets an \code{x}}
\item{metric}{Framework metric}
\item{level_of_analysis}{Level or scope that the metric refers to. Options
are farm, agricultural sector, or population/county/landscape.}
\item{definition}{Longer form definition of the metric, usually pulled
from the source itself.}
\item{resolution}{Smallest geographic scope that the metric can be used at.
This is often county or state, but if spatial data, could also be vector
or the size of raster cells (e.g., 30m).}
\item{updates}{How often the metric is updated from the source}
\item{weighting}{What should be used to weight the variable to treat all
counties equally}
\item{desirable}{If possible, which direction of the metric would represent
an improvement (i.e., what is "good").}
\item{source}{Source of data}
\item{url}{Website where metric can be downloaded from}
\item{notes}{Relevant notes}
\item{status}{Whether metric is fully cleaned, wrangled, etc.}
\item{variable_name}{Unique text identifier for each metric.}
\item{n_states}{Number of states in the Northeast represented by the
metric. Maximum is 9.}
\item{n_counties}{Number of counties in the Northeast represented by the
metric. While the maximum should technically be 218, it could be 217 if the
metric was only collected before 2022, or 228 if it includes data from both
before and after 2022. This is due to Connecticut changing their county
system to a governance region system circa 2022.}
\item{n_years}{Number of unique years represented by the metric. Note that
this is not the range of years. Many metrics are from USDA NASS, which only
collect them every four years.}
\item{latest_year}{Latest year for which data are available}
\item{shorthand_citations}{Literature justifications to add in a table}
}
}
}
\usage{
dp_meta
}
\description{
Metadata associated with \code{data_paper_metrics} specified in
\code{secondary_metrics_revised.xlsx}.
}
\keyword{datasets}
